requirements.txt
aiogram>=3.14.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.45
aiosqlite==0.20.0
pydantic>=2.10.0
pydantic-settings>=2.5.2
openai>=1.109.1
numpy>=2.0.0
faiss-cpu==1.13.2
langchain>=1.2.0
langchain-openai>=1.0.0
pymupdf==1.24.14
python-docx==1.1.2
google-genai>=0.3.0
pydub>=0.25.1
